requirements.txt
python-dotenv==1.0.0
langchain==0.1.17
langchain-openai==0.0.10
openai==0.27.8  # Latest version of OpenAI